Added Give Function and some Chocolate
I added the Give Action so we can give stuff now to the NPCs. It should properly treat it right.

I was on the place for all this so I did some more that will have to be in the same commit, such as adding some early chocolate functionality. Still need to finish it though